requirements.txt
# 第十六届服务外包创新创业大赛国二作品 Demo - Python 依赖
# Python 3.8+

# 深度学习
torch>=1.10.0
torchvision>=0.11.0

# 目标检测 (YOLO / RT-DETR)
ultralytics>=8.0.0

# 图像处理与科学计算
opencv-python>=4.5.0
numpy>=1.19.0
scipy>=1.7.0
Pillow>=8.0.0
PyWavelets>=1.1.1

# 主界面 (明眸视觉)
PySide6>=6.0.0

# 工具
tqdm>=4.60.0
